updated my code

+++ src/Userinput.java
@@ -4,10 +4,12 @@ public class Userinput {
         System.out.println("taking user input");
         Scanner sc = new Scanner(System.in);
         System.out.println("enter the 1st number");
-        int a = sc.nextInt();
+       // int a = sc.nextInt();
+       float a = sc.nextFloat();
         System.out.println("enter the 2nd number");
-        int b = sc.nextInt();
-         int multi = a * b ;
+       // int b = sc.nextInt();
+       float b =sc.nextFloat();
+         float multi = a * b ;
         System.out.println("the multiplication of both number is" + multi);
     }
 }
